+++ src/app/admin/guards/order-resolve.guard.ts
@@ -1,15 +1,22 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, Router } from '@angular/router';
+import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
 
 import { Observable, of } from 'rxjs';
 import { catchError, map, take } from 'rxjs/operators';
 
-import { Order } from 'src/app/orders/models/order.model';
-import { OrderService } from 'src/app/orders/services/order.service';
+import { Store } from '@ngrx/store';
+
+import { Order } from '../../orders/models/order.model';
+import { OrderService } from '../../orders/services/order.service';
+import { AppState } from '../../core/state/app.state';
+import { Go } from '../../core/state/router/router.actions';
 
 @Injectable()
 export class OrderResolveGuard implements Resolve<Order> {
-  constructor(private orderService: OrderService, private router: Router) {}
+  constructor(
+    private orderService: OrderService,
+    private store: Store<AppState>
+  ) {}
 
   resolve(route: ActivatedRouteSnapshot): Observable<Order> {
     if (!route.paramMap.has('orderID')) {
@@ -29,12 +36,12 @@ export class OrderResolveGuard implements Resolve<Order> {
           return order;
         }
 
-        this.router.navigate(['/admin/orders']);
+        this.store.dispatch(new Go({ path: ['/admin/orders'] }));
         return null;
       }),
       take(1),
       catchError(() => {
-        this.router.navigate(['/admin/orders']);
+        this.store.dispatch(new Go({ path: ['/admin/orders'] }));
         return of(null);
       })
     );