+++ src/app/admin/guards/product-resolve.guard.ts
@@ -1,5 +1,10 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, Resolve, Router } from '@angular/router';
+import { ActivatedRouteSnapshot, Resolve } from '@angular/router';
+
+import { Store } from '@ngrx/store';
+
+import { AppState } from '../../core/state/app.state';
+import { Go } from '../../core/state/router/router.actions';
 import { ProductModel } from '../../products/models/product.model';
 import { ProductsService } from '../../products/services';
 
@@ -7,7 +12,7 @@ import { ProductsService } from '../../products/services';
 export class ProductResolveGuard implements Resolve<ProductModel> {
   constructor(
     private productsService: ProductsService,
-    private router: Router
+    private store: Store<AppState>
   ) {}
 
   async resolve(route: ActivatedRouteSnapshot): Promise<ProductModel> {
@@ -21,11 +26,11 @@ export class ProductResolveGuard implements Resolve<ProductModel> {
       if (p) {
         return p;
       } else {
-        this.router.navigate(['/admin/products']);
+        this.store.dispatch(new Go({ path: ['/admin/products'] }));
         return null;
       }
     } catch (e) {
-      this.router.navigate(['/admin/products']);
+      this.store.dispatch(new Go({ path: ['/admin/products'] }));
       return null;
     }
   }