+++ src/app/core/state/core-store.module.ts
@@ -1,0 +1,26 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { EffectsModule } from '@ngrx/effects';
+import { StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
+import { StoreRouterConnectingModule } from '@ngrx/router-store';
+
+import { environment } from '../../../environments/environment';
+import {
+  routerReducers,
+  RouterStateSerializerProvider
+} from './router/router.state';
+import { RouterEffects } from './router/router.effects';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    StoreModule.forRoot(routerReducers),
+    StoreRouterConnectingModule.forRoot(),
+    EffectsModule.forRoot([RouterEffects]),
+    !environment.production ? StoreDevtoolsModule.instrument() : []
+  ],
+  providers: [RouterStateSerializerProvider]
+})
+export class CoreStoreModule {}