+++ changelog.md
@@ -51,3 +51,20 @@
 * Use http client to manage orders. Convert methods in OrderService to use Observable
 * Add timing interceptor for orders requests
 * Use AppSettings service to load service either from local storage, server or set default
+
+## Task 7
+* Add ngrx effects module, ngrx store devtools module.
+* Add get product from url effect and use it on product details component.
+* Add navigation by actions for products
+* Add product selector by url. Use the selector in ProductDetails component.
+* Add router state.
+* Add products state selectors and use them in components.
+* Add products state feature selector.
+* Add product with state.
+* Add @ngrx/store package. Add ProductsState, AppState. Use ProductsState in the products list component.
+* Use ngrx entity for products state.
+* Go back after product edit on Save button.
+* Edit product with effect and store.
+* Get product from database via effect.
+* Get products from database via effect.
+* Modify state through products reducer.